Possible fix for race condition by checking if the file still exists

+++ functions/index.js
@@ -12,9 +12,19 @@ exports.storageFinalize = functions.storage.object().onFinalize((object, context
 });
 
 exports.storageDelete = functions.storage.object().onDelete((object, context) => {
-	// Delete file by name
-	return filesCollection
-		.doc(object.name)
-		.delete()
-		.then(() => console.log('Deleted file information'));
-})
+	// Check if file still exists
+	return admin.storage().bucket().file(object.name).exists().then(exists => {
+		if (exists[0]) {
+			// File exists. This means this was an file update operation
+			// and not a file delete operation.
+			// We can abort the delete mirror to firestore.
+			return false;
+		}
+
+		// Delete file by name
+		return filesCollection
+			.doc(object.name)
+			.delete()
+			.then(() => console.log('Deleted file information'));
+	});
+});
